requirements.txt
numpy==1.25.2
openpyxl==3.1.5
httpx==0.26.0
google-genai==0.3.0
fake-useragent==2.0.3
requests==2.31.0
Pillow==10.0.0